6th commit

+++ server/app.py
@@ -44,7 +44,7 @@ def predict():
         features = [airline, source_city, departure_time, stops, arrival_time, destination_city, travel_class, date_diff]
         prediction = model.predict([features])[0]
 
-        return jsonify({'prediction': round(prediction, 2)})
+        return jsonify({'prediction': float(round(prediction, 2))})
     except KeyError as e:
         return jsonify({'error': f'Missing data for: {e}'}), 400
     except Exception as e:
